pyproject.toml
[project]
name = "vinasltpaymentsservice"
version = "0.1.0"
description = ""
authors = [
    {name = "peyro",email = "peyrovskaaa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "stripe (>=12.2.0,<13.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "fastapi-problem (>=0.11.4,<0.12.0)",
    "aio-pika (>=9.5.7,<10.0.0)",
    "grpcio (>=1.74.0,<2.0.0)",
    "grpcio-tools (>=1.74.0,<2.0.0)",
    "protobuf (>=6.32.0,<7.0.0)",
    "googleapis-common-protos (>=1.70.0,<2.0.0)",
    "grpcio-reflection (>=1.74.0,<2.0.0)",
    "grpcio-health-checking (>=1.74.0,<2.0.0)",
    "authtools @ git+https://github.com/vinasLT/shared_auth.git@v0.1.8",
    "fastapi-pagination (>=0.15.0,<0.16.0)",
    "aiosqlite (>=0.20.0,<0.21.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
